feat update WithField

+++ cmd/protoc-gen-tool/internal/template/router/newProtobuf.go
@@ -2,7 +2,7 @@
  * @Author: Jerry.Yang
  * @Date: 2023-12-12 16:19:17
  * @LastEditors: Jerry.Yang
- * @LastEditTime: 2025-03-31 16:29:29
+ * @LastEditTime: 2025-06-26 16:50:18
  * @Description: new protobuf
  */
 package router
@@ -173,7 +173,7 @@ func (n *NewProtobuf) getTemplate() string {
 	   outputVo, err := {{.FirstRouterName}}.HttpServer.{{.RouterFunc}}(ctx, inputVo)
 	   if err != nil {
 		   qlog.Errorf("{{.RouterNameUp}}Service {{.RouterFunc}} Err : %+v", err)
-		   ctx.JSON(http.StatusBadRequest,&protobuf.{{.OutputRespName}}{RetCode:proto.Int64(-1),RetMsg: proto.String(err.Error())})
+		   ctx.JSON(http.StatusBadRequest,&protobuf.{{.OutputRespName}}{RetCode:proto.Int32(-1),RetMsg: proto.String(err.Error())})
 		   return
 	   }
  
@@ -181,7 +181,7 @@ func (n *NewProtobuf) getTemplate() string {
 	    * @step
 	    * @return
 	   **/
-	   outputVo.RetCode = proto.Int64(0)
+	   outputVo.RetCode = proto.Int32(0)
 	   outputVo.RetMsg = proto.String("")
 	   ctx.JSON(http.StatusOK, outputVo)
    }
